third_party/blink/web_tests/platform/generic/fast/events/drop-generate-user-gesture-expected.txt

Drag the following text into the rectangle:



Drag me

Test that drop event generates user gesture

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.open('about:blank') is defined.
PASS successfullyParsed is true

TEST COMPLETE
